+++ test/xboo-test.t.sol
@@ -6,41 +6,30 @@ import "@openzeppelin/contracts/proxy/ERC1967/ERC1967Proxy.sol";
 import "../contracts/ReaperVaultv1_3.sol";
 import "./abstracts/XbooConstants.t.sol";
 
-
 contract xBooTest is XbooConstants {
-    
-   
     ReaperAutoCompoundXBoov2 XbooStrat;
     ReaperAutoCompoundXBoov2 stratIMPL;
     ReaperVaultv1_3 vault;
     ERC1967Proxy stratProxy;
-    
-    
+
     function setUp() public virtual {
-        vault = new ReaperVaultv1_3(
-            address(Boo),
-            "XBOO Single Stake Vault",
-            "rfXBOO",
-            0,
-            0,
-            type(uint).max
-        );
+        vault = new ReaperVaultv1_3(address(Boo), "XBOO Single Stake Vault", "rfXBOO", 0, 0, type(uint256).max);
         vm.label(address(vault), "boo vault");
-        
+
         stratIMPL = new ReaperAutoCompoundXBoov2();
         vm.label(address(stratIMPL), "strategy Implementation");
         stratProxy = new ERC1967Proxy(
-            address (stratIMPL),
+            address(stratIMPL),
             "" //args
         );
         vm.label(address(stratProxy), "ERC1967 Proxy: Strategy Proxy");
 
         XbooStrat = ReaperAutoCompoundXBoov2(address(stratProxy));
 
-        address[] memory feeRemitters = new address[](2); 
+        address[] memory feeRemitters = new address[](2);
         feeRemitters[0] = address(0xb0C9D5851deF8A2Aac4A23031CA2610f8C3483F9);
         feeRemitters[1] = address(1);
-        address[] memory strategists = new address[](1); 
+        address[] memory strategists = new address[](1);
 
         address[] memory msRoles = new address[](3);
         msRoles[0] = address(0xb0C9D5851deF8A2Aac4A23031CA2610f8C3483F9);
@@ -74,7 +63,7 @@ contract xBooTest is XbooConstants {
 
         vm.label(currentAceLab, "aceLab");
         vm.label(currentMagicats, "magicats");
-        vm.label(WFTM,"WFTM");
+        vm.label(WFTM, "WFTM");
         vm.label(0xa48d959AE2E88f1dAA7D5F611E01908106dE7598, "xBoo");
         vm.label(0xb0C9D5851deF8A2Aac4A23031CA2610f8C3483F9, "admin ms");
     }
@@ -95,30 +84,29 @@ contract xBooTest is XbooConstants {
         uint256 endingBalance = Boo.balanceOf(user1);
         console2.log("ending balance is: %s", endingBalance);
         assertGe(startingBalance, endingBalance);
-       
     }
 
     function testDepositAndHarvestAndSeeYield() public {
         vm.startPrank(user1);
         uint256 startingBalance = Boo.balanceOf(user1);
-        
+
         console2.log("starting balance is : %s", startingBalance);
-        
+
         vault.deposit(startingBalance);
         vm.stopPrank();
-        
+
         setAllocations();
         IAceLab(currentAceLab).massUpdatePools();
-        
+
         console2.log("deposited all boo and recieve %s shares", vault.balanceOf(user1));
         console2.log("strat believes it has a total of %s boo", XbooStrat.balanceOfPool());
-        uint time = uint(block.timestamp);
+        uint256 time = uint256(block.timestamp);
         uint256 apr;
-        uint iterations = 10;
-        for(uint i; i < iterations; i++){
+        uint256 iterations = 10;
+        for (uint256 i; i < iterations; i++) {
             vm.warp(time += 13 hours);
             XbooStrat.harvest();
-            apr = uint(XbooStrat.averageAPRAcrossLastNHarvests(6));
+            apr = uint256(XbooStrat.averageAPRAcrossLastNHarvests(6));
             console.log("APR is : %s", apr);
         }
     }
@@ -127,7 +115,7 @@ contract xBooTest is XbooConstants {
         vm.prank(0xb0C9D5851deF8A2Aac4A23031CA2610f8C3483F9);
         XbooStrat.pause();
         vm.stopPrank();
-        
+
         vm.prank(user1);
         vm.expectRevert(bytes("Pausable: paused"));
         vault.deposit(100 ether);
@@ -136,7 +124,7 @@ contract xBooTest is XbooConstants {
         vm.prank(0xb0C9D5851deF8A2Aac4A23031CA2610f8C3483F9);
         XbooStrat.unpause();
         vm.stopPrank();
-        
+
         vm.prank(user1);
         vault.deposit(100 ether);
         vm.stopPrank();
@@ -159,21 +147,20 @@ contract xBooTest is XbooConstants {
 
         uint256 hecPoolBalanceAfter = XbooStrat.poolXBOOBalance(HEC_ID);
         assertEq(hecPoolBalanceAfter, 0);
-        
+
         uint256 strategyInternalAccountingDepositedAmount = XbooStrat.totalPoolBalance();
         assertEq(strategyInternalAccountingDepositedAmount, 0, "internalAccounting!=0");
         uint256 strategyAcelabDepositedAmount = IAceLab(currentAceLab).balanceOf(address(XbooStrat));
-        assertEq(strategyInternalAccountingDepositedAmount, strategyAcelabDepositedAmount, "internalAcconting!=actualState");
+        assertEq(
+            strategyInternalAccountingDepositedAmount,
+            strategyAcelabDepositedAmount,
+            "internalAcconting!=actualState"
+        );
 
         uint256 endingBalance = XbooStrat.balanceOf();
 
         assertGe(endingBalance, startingBalance);
-
     }
-
-
-
-
 
     /*
     ///////////////////// HELPER FUNCTIONS ////////////////////
@@ -181,19 +168,30 @@ contract xBooTest is XbooConstants {
 
     function setRoutes() public {
         address[] memory hecRoute = new address[](3);
-        hecRoute[0] = HEC; hecRoute[1] = USDC; hecRoute[2] = WFTM; 
+        hecRoute[0] = HEC;
+        hecRoute[1] = USDC;
+        hecRoute[2] = WFTM;
         address[] memory lqdrRoute = new address[](2);
-        lqdrRoute[0] = LQDR; lqdrRoute[1] = WFTM;
+        lqdrRoute[0] = LQDR;
+        lqdrRoute[1] = WFTM;
         address[] memory OrbsRoute = new address[](3);
-        OrbsRoute[0] = ORBS; OrbsRoute[1] = USDC; OrbsRoute[2] = WFTM;
+        OrbsRoute[0] = ORBS;
+        OrbsRoute[1] = USDC;
+        OrbsRoute[2] = WFTM;
         address[] memory xTarotRoute = new address[](2);
-        xTarotRoute[0] = Tarot; xTarotRoute[1] = WFTM;
+        xTarotRoute[0] = Tarot;
+        xTarotRoute[1] = WFTM;
         address[] memory GALCXRoute = new address[](2);
-        GALCXRoute[0] = GALCX; GALCXRoute[1] = WFTM;
+        GALCXRoute[0] = GALCX;
+        GALCXRoute[1] = WFTM;
         address[] memory SDRoute = new address[](3);
-        SDRoute[0] = SD; SDRoute[1] = USDC; SDRoute[2] = WFTM;
+        SDRoute[0] = SD;
+        SDRoute[1] = USDC;
+        SDRoute[2] = WFTM;
         address[] memory singleRoute = new address[](3);
-        singleRoute[0] = SINGLE; singleRoute[1] = USDC; singleRoute[2] = WFTM; 
+        singleRoute[0] = SINGLE;
+        singleRoute[1] = USDC;
+        singleRoute[2] = WFTM;
 
         XbooStrat.setRoute(HEC_ID, hecRoute);
         XbooStrat.setRoute(LQDR_ID, lqdrRoute);
@@ -212,66 +210,58 @@ contract xBooTest is XbooConstants {
         vm.label(SINGLE, "SINGLE");
         vm.label(USDC, "USDC");
         vm.label(uniRouter, "SpookySwap Router");
-
     }
 
-    function setAllocations() public{
-        uint hecAlloc = 10000;
-        uint orbsAlloc = 0;
-        uint galcxAlloc = 0;
-        uint xTarotAlloc = 0;
-        uint lqdrAlloc = 0;
-        uint stratBalance = XbooStrat.balanceOfPool();
-        uint length = IAceLab(currentAceLab).poolLength();
-        uint[] memory idealAmounts = new uint[](length);
-        uint[] memory currentAmounts = new uint[](length);
-        for(uint i ; i < length; i++){
-            uint temp;
-            (temp,,,) = IAceLab(currentAceLab).userInfo(i, address(XbooStrat));
+    function setAllocations() public {
+        uint256 hecAlloc = 10000;
+        uint256 orbsAlloc = 0;
+        uint256 galcxAlloc = 0;
+        uint256 xTarotAlloc = 0;
+        uint256 lqdrAlloc = 0;
+        uint256 stratBalance = XbooStrat.balanceOfPool();
+        uint256 length = IAceLab(currentAceLab).poolLength();
+        uint256[] memory idealAmounts = new uint256[](length);
+        uint256[] memory currentAmounts = new uint256[](length);
+        for (uint256 i; i < length; i++) {
+            uint256 temp;
+            (temp, , , ) = IAceLab(currentAceLab).userInfo(i, address(XbooStrat));
             currentAmounts[i] = temp;
-            if(i == HEC_ID){
+            if (i == HEC_ID) {
                 idealAmounts[i] = (hecAlloc * stratBalance) / 10000;
-            }
-            else if(i == LQDR_ID){
+            } else if (i == LQDR_ID) {
                 idealAmounts[i] = (lqdrAlloc * stratBalance) / 10000;
-            }
-            else if(i == ORBS_ID){
+            } else if (i == ORBS_ID) {
                 idealAmounts[i] = (orbsAlloc * stratBalance) / 10000;
-            }
-            else if(i == xTarot_ID){
+            } else if (i == xTarot_ID) {
                 idealAmounts[i] = (xTarotAlloc * stratBalance) / 10000;
-            }
-            else if(i == GALCX_ID){
+            } else if (i == GALCX_ID) {
                 idealAmounts[i] = (galcxAlloc * stratBalance) / 10000;
-            }
-            else{
+            } else {
                 idealAmounts[i] = 0;
             }
-            
         }
-        uint amtWithdraws;
-        uint amtDeposits;
-        uint[] memory withdrawPoolIds = new uint[](length);
-        uint[] memory withdrawAmounts = new uint[](length);
-        uint[] memory depositPoolIds = new uint[](length);
-        uint[] memory depositAmounts = new uint[](length);
+        uint256 amtWithdraws;
+        uint256 amtDeposits;
+        uint256[] memory withdrawPoolIds = new uint256[](length);
+        uint256[] memory withdrawAmounts = new uint256[](length);
+        uint256[] memory depositPoolIds = new uint256[](length);
+        uint256[] memory depositAmounts = new uint256[](length);
 
-        for(uint i; i < length; i++ ){
-            if(idealAmounts[i] > currentAmounts[i]){
+        for (uint256 i; i < length; i++) {
+            if (idealAmounts[i] > currentAmounts[i]) {
                 //deposit
                 depositPoolIds[amtDeposits] = i;
                 depositAmounts[amtDeposits] = idealAmounts[i] - currentAmounts[i];
-                amtDeposits+=1;
+                amtDeposits += 1;
             }
-            if(idealAmounts[i] < currentAmounts[i]){
-                 //deposit
+            if (idealAmounts[i] < currentAmounts[i]) {
+                //deposit
                 withdrawPoolIds[amtWithdraws] = i;
                 withdrawAmounts[amtWithdraws] = currentAmounts[i] - idealAmounts[i];
-                amtWithdraws+=1;               
+                amtWithdraws += 1;
             }
         }
 
         XbooStrat.setXBooAllocations(withdrawPoolIds, withdrawAmounts, depositPoolIds, depositAmounts);
-
     }
 }
